--- src/cas/ac-montpellier.js
@@ -1,141 +1,0 @@
-const fs = require('fs');
-
-const jsdom = require('jsdom');
-const { JSDOM } = jsdom;
-
-const request = require('../request');
-const util = require('../util');
-
-async function login({ username, password, url })
-{
-    console.log(`Logging in '${username}' for '${url}' using Montpellier CAS`);
-
-    let jar = new jsdom.CookieJar();
-    let dom = await start(url, jar);
-
-    dom.window.document.getElementById('user').value = username;
-    dom.window.document.getElementById('password').value = password;
-
-    dom.window.eval('creerCookie(document.getElementById(\'user\'), document.getElementById(\'password\'));');
-
-    let result = await submitForm({
-        dom,
-        jar
-    });
-
-    await submitForm({
-        dom: result,
-        jar,
-        forLogin: false,
-        followRedirects: false
-    });
-
-    /*dom = await start(url, jar, false);
-    dom = await submitForm({
-        dom,
-        jar,
-        asIs: true
-    });*/
-
-    dom = await getDOM({
-        url: `https://www.environnementnumeriquedetravail.fr/cas/login?service=${encodeURIComponent(url)}&idpId=2`,
-        jar,
-        asIs: true
-    });
-
-    //if (dom.indexOf('ATEN') !== -1)
-    if (dom.indexOf('PRONOTE') === -1)
-    {
-        console.log(`Wrong IDs for '${username}'`);
-        throw 'Mauvais Identifiants';
-    }
-
-    /*console.log(dom);
-    process.exit();*/
-
-    console.log(`Logged in '${username}'`);
-
-    return util.extractStart(dom);
-}
-
-async function start(url, jar, first = true)
-{
-    return await submitForm({
-        dom: await getDOM({
-            url: `https://www.environnementnumeriquedetravail.fr/cas/login?service=${encodeURIComponent(url)}&idpId=2`,
-            jar
-        }),
-        jar,
-        forLogin: first
-    });
-}
-
-function submitForm({ dom, jar, forLogin, followRedirects, asIs })
-{
-    let url = dom.window.document.getElementsByTagName('form')[0].action;
-
-    if (url.indexOf('/') === -1)
-    {
-        url = 'https://famille.ac-montpellier.fr/login/' + url;
-    }
-
-    let params = getParams(dom);
-
-    let data = {
-        url,
-        jar,
-        asIs,
-        followRedirects,
-
-        data: params,
-        method: 'POST'
-    };
-
-    if (forLogin)
-    {
-        data['runScripts'] = true;
-        data['hook'] = (window) => {
-            window.eval(fs.readFileSync('./jsencrypt.min.js') + '; window.JSEncrypt = JSEncrypt;');
-        };
-    }
-
-    return getDOM(data);
-}
-
-function getParams(dom)
-{
-    let params = {};
-
-    Array.prototype.forEach.call(dom.window.document.getElementsByTagName('input'), input => input.name !== '' ? params[input.name] = input.value : '');
-
-    return params;
-}
-
-async function getDOM({ url, jar, method = 'GET', data = '', runScripts, hook = () => {}, followRedirects, asIs })
-{
-    let result = await request.http({
-        url,
-        method,
-        data,
-        jar,
-        followRedirects
-    });
-
-    if (asIs)
-    {
-        return result;
-    }
-
-    if (result.indexOf('<script>$(function() { startup() });</script>') !== -1)
-    {
-        result = result.replace('<script>$(function() { startup() });</script>', '').replace('console.log(user+" "+pwd);', '');
-    }
-
-    return new JSDOM(result, {
-        runScripts: runScripts ? 'dangerously' : 'outside-only',
-        beforeParse(window) { hook(window) },
-        cookieJar: jar
-    });
-}
-
-module.exports = login;
