--- src/cas/ac-reims.js
@@ -1,9 +1,0 @@
-const mbn = require('./mbn');
-
-module.exports = function(params)
-{
-    return mbn(params, {
-        name: 'ac-reims',
-        idp: 'REIMS-ATS_parent_eleve'
-    });
-};
